% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calc_res.R
\name{fct_calc_res}
\alias{fct_calc_res}
\title{Calculate median and confidence interval based on Monte Carlo Simulations}
\usage{
fct_calc_res(.data, .id, .sim, .ci_alpha)
}
\arguments{
\item{.data}{a data frame containing the simulation aggregated results}

\item{.id}{the name of the ID column of the data frame}

\item{.sim}{the name of the column containing the simulated  values}

\item{.ci_alpha}{alpha value for the confidence interval with confidence level = 1 - alpha.}
}
\value{
A tibble with simulation results per category if ID column: estimated mean,
percentage uncertainty, margin of error, lower and upper bound of confidence interval.
}
\description{
Monte Carlo simulations from other package functions generate
a table with one row = one simulation. fct_calc_res() takes the
simulation table and output the simulated variable's median and
its confidence interval for each of the selected category (land
use transition, REDD+ activity or time period).
}
\examples{
## TBD

}
