% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_combine_mcs_C.R
\name{fct_combine_mcs_C}
\alias{fct_combine_mcs_C}
\title{Generate Monte Carlo Simulations and calculate carbon stock for all pools}
\usage{
fct_combine_mcs_C(.c_sub, .c_unit, .c_fraction = NA, .n_iter)
}
\arguments{
\item{.cs}{Subset with one land use from the carbon stock input table for the shiny app (c_stock)}
}
\value{
A data frame with N simulations of each carbon pool and associated factors and the resulting
total carbon stock in tCO2/ha
}
\description{
TBD
}
\examples{
library(mocaredd)
library(readxl)
library(dplyr)
library(ggplot2)

cs <- read_xlsx(system.file("extdata/example1.xlsx", package = "mocaredd"), sheet = "c_stock", na = "NA")
c_sub <- cs |> filter(lu_id == "ev_wet_closed")
res <- fct_combine_mcs_cstock(.n_iter = 10000, .c_sub = c_sub, .c_unit = "C")
res

}
