% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_combine_mcs_cpools.R
\name{fct_combine_mcs_cpools}
\alias{fct_combine_mcs_cpools}
\title{Generate Monte Carlo Simulations and calculate carbon stock for all pools}
\usage{
fct_combine_mcs_cpools(.c_sub, .usr)
}
\arguments{
\item{.c_sub}{Subset with one land use from the carbon stock input table for the shiny app (c_stocks).}

\item{.usr}{User inputs' table for the shiny app (user_inputs). Contains the number
of iterations of the MCS, carbon fraction if needed and if truncated PDFs
should be used when necessary.}
}
\value{
A data frame with N simulations of each carbon pool and associated factors and the resulting
total carbon stock in tCO2/ha.
}
\description{
TBD
}
\examples{
library(dplyr)
library(readxl)
library(mocaredd)

path <- system.file("extdata/example1-4pools.xlsx", package = "mocaredd")

usr <- read_xlsx(path = path, sheet = "user_inputs", na = "NA")
cs <- read_xlsx(path = path, sheet = "c_stocks", na = "NA")

cs_clean <- cs |> filter(!(is.na(c_value) & is.na(c_pdf_a)))
c_sub <- cs_clean |> filter(lu_id == "ev_wet_closed")

res <- fct_combine_mcs_cpools(.c_sub = c_sub, .usr = usr)

hist(res$C_all)
round(median(res$C_all))

}
