% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_combine_mcs_cstock.R
\name{fct_combine_mcs_cstock}
\alias{fct_combine_mcs_cstock}
\title{Generate and propagate Monte Carlo Simulations based on a template input file.}
\usage{
fct_combine_mcs_cstock(.ad, .cs, .usr)
}
\arguments{
\item{.ad}{Activity Data input table for the shiny app (AD_lu_transitions)}

\item{.cs}{Carbon Stock input table for the shiny app (c_stocks)}

\item{.usr}{User inputs' table for the shiny app (user_inputs). Contains the number
of iterations of the MCS, carbon fraction if needed and if truncated PDFs
should be used when necessary.}
}
\value{
A data frame with Monte Carlo simulations of CO2 emissions for each land use
transition, REDD+ activity or emission reductions level.
}
\description{
TBD
}
\examples{
library(mocaredd)
library(readxl)
library(dplyr)

path <- system.file("extdata/example2-with-sims.xlsx", package = "mocaredd")

cs <- read_xlsx(path = path, sheet = "c_stocks", na = "NA")
ad <- read_xlsx(path = path, sheet = "AD_lu_transitions", na = "NA")
usr <- read_xlsx(path = path, sheet = "user_inputs", na = "NA")

res <- fct_combine_mcs_cstock(.ad = ad, .cs = cs, .usr = usr)
res |> filter(sim_no == 1)

}
