% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_combine_mcs_E.R
\name{fct_combine_mcs_E}
\alias{fct_combine_mcs_E}
\title{Generate and propagate Monte Carlo Simulations based on a template input file.}
\usage{
fct_combine_mcs_E(.ad, .cs, .init, .usr)
}
\arguments{
\item{.ad}{Activity Data input table for the shiny app (AD_lu_transitions)}

\item{.cs}{Carbon Stock input table for the shiny app (c_stock)}
}
\value{
A list of dataframes with Monte Carlo simulations for input variables, REDD+ activities CO2 emissions
and removals and emission reductions levels.
}
\description{
TBD
}
\examples{
library(mocaredd)
library(readxl)
library(dplyr)
library(ggplot2)

cs <- read_xlsx(system.file("extdata/example1.xlsx", package = "mocaredd"), sheet = "c_stock", na = "NA")
ad <- read_xlsx(system.file("extdata/example1.xlsx", package = "mocaredd"), sheet = "AD_lu_transitions", na = "NA")

res <- fct_combine_mcs_E(.ad = ad, .cs = cs)

## ADD HISTOGRAM FOR ONE SET OF SIMULATIONS

}
