% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_combine_mcs_P.R
\name{fct_combine_mcs_P}
\alias{fct_combine_mcs_P}
\title{Combine MCS of emissions to a defined time period}
\usage{
fct_combine_mcs_P(.data, .time, .period_type, .ad_annual)
}
\arguments{
\item{.data}{a data frame containing the simulations}

\item{.time}{the 'time' table from the tool input file (see template)}

\item{.period_type}{"reference" or "monitoring"}

\item{.ad_annual}{TRUE or FALSE, is the activity data annualized or not.}
}
\value{
A tibble with simulations at the final estimate per type of period.
}
\description{
Depending on how the period is defined and if the data are annualized
or not, calculate the Emission Level for a reference or monitoring period
for each simulation.
}
\examples{
## TBD

}
