% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_histogram.R
\name{fct_histogram}
\alias{fct_histogram}
\title{Generate histogram to show the simulations at any stage.}
\usage{
fct_histogram(.data, .res, .id, .value, .value_type, ...)
}
\arguments{
\item{.data}{data frame of simulated data. Output of fct_combine_mcs_E() or fct_combine_mcs_P().}

\item{.res}{aggregated results of a simulation. Output of fct_calc_res()}

\item{.id}{ID column of .data and .res inputs.}

\item{.value}{simulated values column in .data.}

\item{.value_type}{"E" or "ER" if .data represents emissions or emission reductions respectively.}

\item{...}{filter query, passed to \code{\link[dplyr:filter]{dplyr::filter()}}.}
}
\value{
a ggplot or a named list of ggplots
}
\description{
Takes simulation data from fct_combine_mcs_E() or fct_combine_mcs_P() and
aggregated results from fct_calc_res() and generate an histogram showing the
PDF density, modeled normal distribution, simulations' median and confidence
interval.
}
