% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_make_formula.R
\name{fct_make_formula}
\alias{fct_make_formula}
\title{Check that the app's the input data has matching land uses and correct carbon pools and associated factors.}
\usage{
fct_make_formula(.c_check, .c_unit)
}
\arguments{
\item{.c_check}{output of fct_check_pool()}

\item{.c_unit}{"DM" or "C",}
}
\value{
A character value with the formula for calculating total carbon stock for a specific land use.
}
\description{
The expected input carbon stock table for the tool contains carbon values of
different carbon pools coded:
- AGB for aboveground biomass,
- BGB or RS (Root-to-Shoot ratio) for the belowground biomass,
- DW for deadwood,
- LI for litter and
- SOC for soil organic carbon.
All pools can be expressed in tons of Carbon (C) or CO2.
AGB and BGB can also be expressed in ton of dry matter (DM), then a carbon fraction (CF) is also needed.
}
\examples{
library(mocaredd)
library(readxl)
library(dplyr)

cs    <- read_xlsx(system.file("extdata/example1.xlsx", package = "mocaredd"), sheet = "c_stock", na = "NA")
c_lu  <- cs |> filter(lu_id == "dg_ev_wet_closed")

c_check <- fct_check_pool(.c_lu = c_lu, .c_unit = "C", .c_fraction = NA)

fct_make_formula(.c_check = c_check, .c_unit = "C")

}
