% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_make_mcs.R
\name{fct_make_mcs}
\alias{fct_make_mcs}
\title{Generate Monte Carlo Simulations for a given PDF}
\usage{
fct_make_mcs(
  .n_iter = 10000,
  .pdf,
  .mean = NA,
  .se = NA,
  .params = NULL,
  .trunc = FALSE
)
}
\arguments{
\item{.n_iter}{Number of iterations of the Monte Carlo simulations}

\item{.pdf}{Probability Distribution Function name, supported distributions so far: "normal", "beta".}

\item{.mean}{Mean value if PDF parameters are mean and sd}

\item{.se}{Standard deviation of the mean value if PDF parameters are mean and sd}

\item{.params}{Vector pf parameters for other PDFs (NOT IMPLEMENTED YET)}

\item{.trunc}{TRUE or FALSE, should simulations be truncated to allow only values
above 0 where negative values are impossible}
}
\value{
a vector of N simulation with N, the number of iterations (.n_iter)
}
\description{
TBD
}
\examples{
library(mocaredd)

tt <- fct_make_mcs(.pdf = "normal", .mean = 0, .se = 1)
hist(tt)

}
