% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_round.R
\name{fct_round}
\alias{fct_round}
\title{Reproduce round function from MS Excel}
\usage{
fct_round(.num, .digits)
}
\arguments{
\item{.num}{number to be rounded}

\item{.digits}{Number of digits to be passed to round()}
}
\value{
a rounded numerical value
}
\description{
See: https://stackoverflow.com/questions/59392686/r-how-to-define-round-function-like-in-excel
}
