% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_run_mocaredd.R
\name{shiny_run_mocaredd}
\alias{shiny_run_mocaredd}
\title{Run Monte Carlo for REDD+ uncertainty's shiny app}
\usage{
shiny_run_mocaredd(...)
}
\arguments{
\item{...}{arguments to pass to shinyApp}
}
\description{
Starts a R Shiny application that guide users to fill in context and
land use changes for forest related greenhouse gas emissions and removals.
Inputs include distribution parameters for various distribution then the app
returns error propagation and sensitivity analysis of the overall uncertainty
for GHG estimates.
}
\examples{
if (interactive()) {

shiny_run_mocaredd()

}

}
