% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_data2.R
\name{fct_check_data2}
\alias{fct_check_data2}
\title{Check that the app's input data meet the template criteria.}
\usage{
fct_check_data2(.usr, .time, .ad, .cs)
}
\arguments{
\item{.usr}{User inputs' table for the shiny app (user_inputs)}

\item{.time}{the 'time' table from the tool input file (see template)}

\item{.ad}{Activity Data input table for the shiny app (AD_lu_transitions)}

\item{.cs}{Carbon Stock input table for the shiny app (c_stocks)}
}
\value{
A dataframe with TRUE or FALSE (TRUE if each check passes), and broad error locations
if FALSE.
}
\description{
Tests 6 types of conformity with the app template to avoid code break during
analysis: (1) column names, (2) tables' size, (3) data types,
(4) categories names, (5) unique IDs and (6) matching key variables between tables.
}
\examples{
library(mocaredd)
library(readxl)

path <- system.file("extdata/example1-4pools.xlsx", package = "mocaredd")

cs <- read_xlsx(path = path, sheet = "c_stocks", na = "NA")
ad <- read_xlsx(path = path, sheet = "AD_lu_transitions", na = "NA")
usr <- read_xlsx(path = path, sheet = "user_inputs", na = "NA")
time <- read_xlsx(path = path, sheet = "time_periods", na = "NA")

fct_check_data2(.ad = ad, .cs = cs, .usr = usr, .time = time)

}
