% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_check_pool.R
\name{fct_check_pool}
\alias{fct_check_pool}
\title{Check which Carbon pool and what units are in the carbon stock data.}
\usage{
fct_check_pool(.c_lu, .c_unit, .c_fraction)
}
\arguments{
\item{.c_lu}{Subset of the Carbon Stock input table (c_stock) for a unique land use}

\item{.c_unit}{Unit used to report AGB (and BGB if included), from the shiny app or template user inputs.}

\item{.c_fraction}{Value of carbon fraction, only if dry matter reported}
}
\value{
A dataframe with TRUE or FALSE for each pool and factor in the app's carbon stock input data for a given land use.
}
\description{
The expected input carbon stock table for the tool contains carbon values of
different carbon pools coded:
- AGB for aboveground biomass,
- BGB or RS (Root-to-Shoot ratio) for the belowground biomass,
- DW for deadwood,
- LI for litter and
- SOC for soil organic carbon.
All pools can be expressed in tons of Carbon (C) or CO2.
AGB and BGB can also be expressed in ton of dry matter (DM), then a carbon fraction (CF) is also needed.
}
\examples{
library(mocaredd)
library(readxl)
library(dplyr)

path <- system.file("extdata/example1-4pools.xlsx", package = "mocaredd")

cs <- read_xlsx(path = path, sheet = "c_stocks", na = "NA")

c_lu <- cs |> filter(lu_id == "ev_wet_closed")

fct_check_pool(.c_lu = c_lu, .c_unit = "C", .c_fraction = NA)

}
