% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_combine_mcs_E.R
\name{fct_combine_mcs_E}
\alias{fct_combine_mcs_E}
\title{Generate and propagate Monte Carlo Simulations based on a template input file.}
\usage{
fct_combine_mcs_E(.ad, .cs, .usr)
}
\arguments{
\item{.ad}{Activity Data input table for the shiny app (AD_lu_transitions)}

\item{.cs}{Carbon Stock input table for the shiny app (c_stocks)}

\item{.usr}{User inputs' table for the shiny app (user_inputs). Contains the number
of iterations of the MCS, carbon fraction if needed and if truncated PDFs
should be used when necessary.}
}
\value{
A list of dataframes with Monte Carlo simulations for input variables, REDD+ activities CO2 emissions
and removals and emission reductions levels.
}
\description{
TBD
}
\examples{
library(readxl)
library(dplyr)
library(mocaredd)

cs <- read_xlsx(
  system.file("extdata/example1.xlsx", package = "mocaredd"),
  sheet = "c_stocks",
  na = "NA"
  )
ad <- read_xlsx(
  system.file("extdata/example1.xlsx", package = "mocaredd"),
  sheet = "AD_lu_transitions",
  na = "NA"
  )
usr <- read_xlsx(
  system.file("extdata/example1.xlsx", package = "mocaredd"),
  sheet = "user_inputs",
  na = "NA"
  )

cs_clean <- cs |> filter(!is.na(c_value) | !is.na(c_pdf_a))

res <- fct_combine_mcs_E(.ad = ad, .cs = cs_clean, .usr = usr)
hist(res$E_sim)
round(median(res$E_sim))

}
