% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_make_formula.R
\name{fct_make_formula}
\alias{fct_make_formula}
\title{Check that the app's the input data has matching land uses and correct carbon pools and associated factors.}
\usage{
fct_make_formula(.c_el, .c_unit)
}
\arguments{
\item{.c_el}{Vector of carbon elements, inc. "A::}

\item{.c_unit}{"DM" or "C",}
}
\value{
A character value with the formula for calculating total carbon stock.
}
\description{
The expected input carbon stock table for the tool contains carbon values of
different carbon pools coded:
- AGB for aboveground biomass,
- BGB or RS (Root-to-Shoot ratio) for the belowground biomass,
- DW for deadwood,
- LI for litter and
- SOC for soil organic carbon.
All pools can be expressed in tons of carbon (C), except AGB and BGB
which can also be expressed in ton of dry matter (DM) if a carbon fraction
'CF' is provided.
}
\examples{
library(mocaredd)

c_el <- c("AGB", "RS", "DW")

fct_make_formula(.c_el = c_el, .c_unit = "DM")

}
